Fix error when porting from rstatic
For some reason, at some point I named this function
distribution_static__cmt(), while the equovalent function in
shellfu-bash-rstatic is called rstatic__hint().

That led to this typo when porting newer rstatic changes back to this
library.  (Which was done at the same time as porting it from Red Hat
Internal repo -- for that reason we can't see that in git log.)

The fix is trivial, just use the correct name.